bug 260 fix  * changed visibility of snapshots so that they are not publicly available to be viewed used by any admin     * util.py:        * added:           * has_permission() method that uses our new cascading permissions model where users can see their own and below, with admin being able to see all     * glance_ops_v2.py:        * updated:           * import_image() now tacks on user_id and stores user_level in tags[] upon import (to help with cascading permissions)           * list_images() now checks if a user can see the images based on cascading permissions or if the image is set to public visibility        * added:           * update_image() method that updates metadata about an image, please read the documentation for that method, we are not checking if the property parameter is a valid metadata property as there are a large number with a larger number of acceptable values for each parameter, currently used to set visiblity and tags[]     * server_action.py:        * updated:           * create_instance_snapshot() now defaults to setting snapshots to visibility of private and to also stores user_level in tags[] (to help with cascading permissions)
currently, snapshots are visible in the instance detail view based on the older permission model, admins see all, pu's see everything in a project and users see only their own.  this will be looked at later, perhaps 3.0, but this bug specifically talks about snapshots being visible across the cloud, which has been resolved.

please note that these visibility issues are cropping up because we have changed to creating a project admin with each project which was not how our older code was designed.  eventually, most of the code will have to be looked over and modified.