large third party authentication refactor
in preparation for ldap, much of the existing third party authentication code was found to be shibboleth agnostic, and therefore was moved to the appropriate place within the code structure.  the basic methods of adding users, building projects and authenticating will be identical for shibboleth and ldap users.  all code referencing the old structure has been updated